RavenDB-18018 - addressed PR comments
