+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/ShardedStatsHandlerProcessorForGetStudioFooterStats.cs
@@ -1,11 +1,14 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using JetBrains.Annotations;
+using Raven.Client.Http;
 using Raven.Server.Documents.Handlers.Processors;
-using Raven.Server.Documents.Sharding;
+using Raven.Server.Documents.Operations;
+using Raven.Server.Documents.Sharding.Operations;
 using Raven.Server.Documents.Studio;
 using Raven.Server.ServerWide.Context;
 
-namespace Raven.Server.Documents.ShardedHandlers.Processors
+namespace Raven.Server.Documents.Sharding.Handlers.Processors
 {
     internal class ShardedStatsHandlerProcessorForGetStudioFooterStats : AbstractStatsHandlerProcessorForGetStudioFooterStats<ShardedRequestHandler, TransactionOperationContext>
     {
@@ -13,13 +16,32 @@ namespace Raven.Server.Documents.ShardedHandlers.Processors
         {
         }
         
-        protected override async ValueTask<FooterStatistics> GetFooterStatistics()
+        protected override async ValueTask<FooterStatistics> GetFooterStatisticsAsync()
         {
-            var op = new ShardedStudioStatsHandler.ShardedGetStudioFooterStatsOperation();
+            var op = new ShardedGetStudioFooterStatsOperation();
             var stats = await RequestHandler.ShardExecutor.ExecuteParallelForAllAsync(op);
             stats.CountOfIndexes = RequestHandler.ShardedContext.DatabaseRecord.Indexes.Count;
             //missing CountOfStaleIndexes and CountOfIndexingErrors fields
             return stats;
+        }
+
+        private readonly struct ShardedGetStudioFooterStatsOperation : IShardedOperation<FooterStatistics>
+        {
+            public FooterStatistics Combine(Memory<FooterStatistics> results)
+            {
+                var span = results.Span;
+
+                var combined = new FooterStatistics();
+
+                foreach (var stats in span)
+                {
+                    combined.CountOfDocuments += stats.CountOfDocuments;
+                }
+
+                return combined;
+            }
+
+            public RavenCommand<FooterStatistics> CreateCommandForShard(int shard) => new GetStudioFooterStatisticsOperation.GetStudioFooterStatisticsCommand();
         }
     }
 }
