+++ test/SlowTests/SlowTests/StudioTests.cs
@@ -16,7 +16,7 @@ namespace SlowTests.SlowTests
         }
 
         [Theory]
-        [RavenData(DatabaseMode = RavenDatabaseMode.Single)]
+        [RavenData(DatabaseMode = RavenDatabaseMode.All)]
         public async Task CanGetStudioFooterStatistics(Options options)
         {
             using (var store = GetDocumentStore(options))
@@ -34,41 +34,23 @@ namespace SlowTests.SlowTests
                 Assert.NotNull(stats);
                 Assert.Equal(1, stats.CountOfIndexes);
                 Assert.Equal(2, stats.CountOfDocuments);
-                Assert.Equal(0, stats.CountOfIndexingErrors);
-                Assert.Equal(0, stats.CountOfStaleIndexes);
-            }
-        }
 
-        [Theory]
-        [RavenData(DatabaseMode = RavenDatabaseMode.Sharded)]
-        public async Task ShardedCanGetStudioFooterStatistics(Options options)
-        {
-            //TODO: can unify with non sharded mode after all FooterStatistics fields are decided
-            using (var store = GetDocumentStore(options))
-            {
-                using (var session = store.OpenSession())
+                if (options.DatabaseMode == RavenDatabaseMode.Single)
                 {
-                    session.Store(new User { Name = "SomeDoc" });
-                    session.SaveChanges();
+                    Assert.Equal(0, stats.CountOfIndexingErrors);
+                    Assert.Equal(0, stats.CountOfStaleIndexes);
                 }
-                await store.ExecuteIndexAsync(new AllowedUsers());
-                WaitForIndexing(store, sharded: true);
-
-                var stats = store.Maintenance.Send(new GetStudioFooterStatisticsOperation());
-
-                Assert.NotNull(stats);
-                Assert.Equal(1, stats.CountOfIndexes);
-                Assert.Equal(2, stats.CountOfDocuments);
-                Assert.Equal(null, stats.CountOfIndexingErrors);
-                Assert.Equal(null, stats.CountOfStaleIndexes);
+                else
+                {
+                    Assert.Equal(null, stats.CountOfIndexingErrors);
+                    Assert.Equal(null, stats.CountOfStaleIndexes);
+                }
             }
         }
-
+        
         private class User
         {
             public string Name;
-
-            public bool Banned;
         }
 
         private class AllowedUsers : AbstractIndexCreationTask<User>
@@ -78,8 +60,7 @@ namespace SlowTests.SlowTests
             public AllowedUsers()
             {
                 Map = users => from user in users
-                               where user.Banned == false
-                               select new
+                    select new
                                {
                                    user.Name
                                };
